Переписанные вариант. Версия 0.1

+++ 3 Семестр/Операционные системы/Lab_2/err_handle.cpp
@@ -5,15 +5,15 @@
 #include <stdlib.h>
 #include <arpa/inet.h>
 
-#include "errhandle.h"
+#include "err_handle.h"
 
-int Socket (int domain, int type, int protocol){
+short Socket (int domain, int type, int protocol){
     int res = socket(domain, type, protocol);
     if (res == -1){
         perror("Create socket failed");
         exit(EXIT_FAILURE);
     }
-    return res;
+    return (short)res;
 }
 
 void Bind(int sockfd, const struct sockaddr *addr, socklen_t addrlen) {
@@ -32,21 +32,21 @@ void Listen(int sockfd, int backlog) {
     }
 }
 
-int Accept(int sockfd, struct sockaddr *addr, socklen_t *addrlen) {
+short Accept(int sockfd, struct sockaddr *addr, socklen_t *addrlen) {
     int res = accept(sockfd, addr, addrlen);
     if (res == -1) {
         perror("accept failed");
         exit(EXIT_FAILURE);
     }
-    return res;
+    return (short)res;
 }
 
 void Inet_pton(int af, const char *src, void *dst) {
     int res = inet_pton(af, src, dst);
     if (res == 0) {
         printf("inet_pton failed: src does not contain a character"
-            " string representing a valid network address in the specified"
-            " address family\n");
+               " string representing a valid network address in the specified"
+               " address family\n");
         exit(EXIT_FAILURE);
     }
     if (res == -1) {
